pyproject.toml
[tool.poetry]
name = "rag_engine"
version = "0.1.0"
description = "Modular Retrieval-Augmented Generation (RAG) engine."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
typer = "^0.12.3"
fastapi = "^0.110.0"
streamlit = "^1.35.0"
gradio = "^4.25.0"
pydantic = "^2.7.0"
pyyaml = "^6.0.0"

[tool.poetry.scripts]
rag-engine = "rag_engine.__main__:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
